+++ src/main/java/com/example/dao/impl/UserDaoImpl.java
@@ -37,7 +37,7 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User getById(long id) {
+    public User getById(int id) {
         return entityManager.find(User.class, id);
     }
